Move item metadata fetch into util file in FileProviderUIExt
